Fixed conversions to long in estimate_memory.cpp; added copyright notice.

+++ core/estimate_memory.cpp
@@ -1,5 +1,33 @@
 // Code for estimating how much memory will be needed by imfit
 
+// Reminder: the correct way to cast multiplication of int variables when
+// assigning to a long variable is:
+//    int nCols, nRows;
+//    long nPix;
+//    ...
+//    nPix = (long)nCols * (long)nRows;
+//
+// (Otherwise, what will happen (with GCC 5, at least) is that the multiplication
+// will be done using int values, will overflow, and the overflow value will be
+// stored in the long variable...
+
+// Copyright 2015--2016 by Peter Erwin.
+// 
+// This file is part of Imfit.
+// 
+// Imfit is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free
+// Software Foundation, either version 3 of the License, or (at your
+// option) any later version.
+// 
+// Imfit is distributed in the hope that it will be useful, but WITHOUT ANY
+// WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+// for more details.
+// 
+// You should have received a copy of the GNU General Public License along
+// with Imfit.  If not, see <http://www.gnu.org/licenses/>.
+
 #include <math.h>
 #include <stdio.h>
 
@@ -24,17 +52,17 @@ long EstimateConvolverMemory( const int nModel_cols, const int nModel_rows,
   long  nBytesNeeded = 0;
   int  nCols_padded, nRows_padded, nCols_padded_trimmed;
 
-  nBytesNeeded += (long)(nPSF_cols * nPSF_rows);   // allocated outside
+  nBytesNeeded += (long)nPSF_cols * (long)nPSF_rows;   // allocated outside
   // Convolver stuff
   nCols_padded = nModel_cols + nPSF_cols - 1;
   nRows_padded = nModel_rows + nPSF_rows - 1;
   nCols_padded_trimmed = (int)(floor(nCols_padded/2)) + 1;   // reduced size of r2c/c2r complex array
-  nPaddedPixels = (long)(nCols_padded * nRows_padded);
-  nPaddedPixels_cmplx = (long)(nCols_padded_trimmed * nRows_padded);
+  nPaddedPixels = (long)nCols_padded * (long)nRows_padded;
+  nPaddedPixels_cmplx = (long)nCols_padded_trimmed * (long)nRows_padded;
   // 3 double-precision arrays allocated in Convolver::DoFullSetup:
-  nBytesNeeded += (long)(3 * nPaddedPixels * DOUBLE_SIZE);
+  nBytesNeeded += 3 * nPaddedPixels * DOUBLE_SIZE;
   // 3 fftw_complex arrays allocated in Convolver::DoFullSetup:
-  nBytesNeeded += (long)(3 * nPaddedPixels_cmplx * FFTW_SIZE);
+  nBytesNeeded += 3 * nPaddedPixels_cmplx * FFTW_SIZE;
   
   return nBytesNeeded;
 }
@@ -53,8 +81,8 @@ long EstimateMemoryUse( int nData_cols, int nData_rows, int nPSF_cols, int nPSF_
   int  nCols_padded = 0;
   int  nRows_padded = 0;
   int  nCols_padded_trimmed = 0;
-  long  nDataPixels = (long)(nData_cols * nData_rows);
-  long  dataSize = (long)(nDataPixels * DOUBLE_SIZE);
+  long  nDataPixels = (long)nData_cols * (long)nData_rows;
+  long  dataSize = nDataPixels * DOUBLE_SIZE;
   int  nModel_rows, nModel_cols;
   int  nOversampModel_cols, nOversampModel_rows;
   long  nModelPixels = 0;
@@ -66,16 +94,16 @@ long EstimateMemoryUse( int nData_cols, int nData_rows, int nPSF_cols, int nPSF_
     // we're doing PSF convolution, so model image will be larger
     nModel_cols = nData_cols + 2*nPSF_cols;
     nModel_rows = nData_rows + 2*nPSF_rows;
-    nModelPixels = (long)(nModel_cols * nModel_rows);
+    nModelPixels = (long)nModel_cols * (long)nModel_rows;
     // memory used by Convolver object
     nBytesNeeded += EstimateConvolverMemory(nModel_cols, nModel_rows, nPSF_cols, nPSF_rows);
     // possible extra memory used by convolution with oversampled PSF
     if (nPSF_osamp_cols > 0) {
       nOversampModel_cols = deltaX_osamp*oversampleScale + 2*nPSF_osamp_cols;
       nOversampModel_rows = deltaY_osamp*oversampleScale + 2*nPSF_osamp_rows;
-      nOversampModelPixels = (long)(nOversampModel_cols * nOversampModel_rows);
+      nOversampModelPixels = (long)nOversampModel_cols * (long)nOversampModel_rows;
       // memory for oversampled model image
-      nBytesNeeded += (long)(nOversampModelPixels * DOUBLE_SIZE);
+      nBytesNeeded += nOversampModelPixels * DOUBLE_SIZE;
       // memory used by Convolver object for oversampled convolution
       nBytesNeeded += EstimateConvolverMemory(nOversampModel_cols, nOversampModel_rows, 
       											nPSF_osamp_cols, nPSF_osamp_rows);
@@ -83,7 +111,7 @@ long EstimateMemoryUse( int nData_cols, int nData_rows, int nPSF_cols, int nPSF_
   }
   else
     nModelPixels = nDataPixels;
-  long  modelSize = (long)(nModelPixels * DOUBLE_SIZE); 
+  long  modelSize = nModelPixels * DOUBLE_SIZE; 
   // the following are always allocated
   nBytesNeeded += 3*modelSize;   // modelVector, weightVector, maskVector
   // possible allocations, depending on type of fit and/or outputs requested
